Introduced a new set minus operator, addressed some of the initial confusion with old acc_tri2.m code
Couldn't get the correct code for growing the left lane safe set and struggled with the acc_tri.m code for a long time. Ultimately discovered what caused the 'overlaps' error that I was seeing and created a new set minus operator to address it, but shortly before doing this the correct code for growing the left lane safe set was retrieved. Saving this information for safekeeping.